Clean up the API
Cervus is built with modularity in mind.  This change codifies this
design principle by making it easy to only use select parts of Cervus.

The primary way of using Cervus is by directly importing its modules:

    import { Game } from 'cervus/core';
    import { Box, Plane } from 'cervus/shapes';
    import { basic } from 'cervus/materials';
    import { vec3, mat4 } from 'cervus/math';

When using the built versions of Cervus from the dist/ folder, the same
namespace hierarchy is exported on the Cervus global:

    const game = new Cervus.core.Game();
    const box = new Cervus.shapes.Box({
      material: Cervus.materials.phong
    });